+++ jaxlib/py_values.h
@@ -153,7 +153,11 @@ struct DevicePutInfo {
   static std::unordered_map<std::string, int64_t> GetInfo();
 };
 
-// Tells the C++ code about the Python type LiteralArray.
+// Tells the C++ code about the Python types LiteralInt, LiteralFloat,
+// LiteralComplex, and LiteralArray.
+void SetLiteralIntType(nanobind::object t);
+void SetLiteralFloatType(nanobind::object t);
+void SetLiteralComplexType(nanobind::object t);
 void SetLiteralArrayType(nanobind::object t);
 
 }  // namespace jax