+++ tests/api_test.py
@@ -52,6 +52,7 @@ from jax._src import config
 from jax._src import core
 from jax._src import dispatch
 from jax._src import linear_util as lu
+from jax._src import literals
 from jax._src import test_util as jtu
 from jax._src import xla_bridge
 from jax._src import debugging
@@ -2086,6 +2087,39 @@ class APITest(jtu.JaxTestCase):
 
     tree_util.register_pytree_node(CustomNode, lambda x: (x, None), unflatten)
     jax.device_put(CustomNode([0.1]))
+
+  def test_device_put_literals(self):
+    self.assertEqual(
+        np.dtype(np.int32),
+        jax.device_put(literals.LiteralInt(1, np.dtype(np.int32))).dtype)
+    self.assertEqual(
+        np.dtype(np.int64),
+        jax.device_put(literals.LiteralInt(1, np.dtype(np.int64))).dtype)
+    self.assertEqual(
+        np.dtype(np.float32),
+        jax.device_put(literals.LiteralFloat(1, np.dtype(np.float32))).dtype)
+    self.assertEqual(
+        np.dtype(np.float64),
+        jax.device_put(literals.LiteralFloat(1, np.dtype(np.float64))).dtype)
+    self.assertEqual(
+        np.dtype(np.complex64),
+        jax.device_put(literals.LiteralComplex(
+            1,np.dtype(np.complex64))).dtype)
+    if jtu.device_under_test() != "tpu":
+      # The TPU compiler does not support complex128.
+      self.assertEqual(
+          np.dtype(np.complex128),
+          jax.device_put(literals.LiteralComplex(
+              1, np.dtype(np.complex128))).dtype)
+    self.assertEqual(
+        np.dtype(np.int32),
+        jax.device_put(literals.LiteralArray(np.array([1], dtype=np.int32),
+                                             weak_type=False)).dtype)
+    self.assertEqual(
+        np.dtype(np.int64),
+        jax.device_put(literals.LiteralArray(np.array([1], dtype=np.int64),
+                                             weak_type=False)).dtype)
+
 
   def test_vmap_inconsistent_sizes_constructs_proper_error_message(self):
     def f(x1, x2, g):